+++ src/main/java/com/moim/backend/domain/user/response/kakao/KakaoTokenResponse.java
@@ -1,4 +1,4 @@
-package com.moim.backend.domain.user.response;
+package com.moim.backend.domain.user.response.kakao;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
@@ -8,17 +8,20 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class NaverTokenResponse {
+public class KakaoTokenResponse {
+
     @JsonProperty("access_token")
     private String accessToken;
-    @JsonProperty("refresh_token")
-    private String refreshToken;
     @JsonProperty("token_type")
     private String tokenType;
+    @JsonProperty("refresh_token")
+    private String refreshToken;
+    @JsonProperty("id_token")
+    private String idToken;
     @JsonProperty("expires_in")
-    private String expiresIn;
-    @JsonProperty("error")
-    private String error;
-    @JsonProperty("error_description")
-    private String errorDescription;
+    private int expiresIn;
+    private String scope;
+    @JsonProperty("refresh_token_expires_in")
+    private int refreshTokenExpiresIn;
+
 }
