+++ src/main/java/com/moim/backend/domain/user/service/GoogleLoginService.java
@@ -3,8 +3,8 @@ package com.moim.backend.domain.user.service;
 import com.moim.backend.domain.user.config.GoogleProperties;
 import com.moim.backend.domain.user.config.Platform;
 import com.moim.backend.domain.user.entity.Users;
-import com.moim.backend.domain.user.response.GoogleTokenResponse;
-import com.moim.backend.domain.user.response.GoogleUserResponse;
+import com.moim.backend.domain.user.response.google.GoogleTokenResponse;
+import com.moim.backend.domain.user.response.google.GoogleUserResponse;
 import com.moim.backend.global.common.exception.CustomException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
