+++ src/main/java/com/moim/backend/domain/user/service/KakaoLoginService.java
@@ -3,8 +3,8 @@ package com.moim.backend.domain.user.service;
 import com.moim.backend.domain.user.config.KakaoProperties;
 import com.moim.backend.domain.user.config.Platform;
 import com.moim.backend.domain.user.entity.Users;
-import com.moim.backend.domain.user.response.KakaoTokenResponse;
-import com.moim.backend.domain.user.response.KakaoUserResponse;
+import com.moim.backend.domain.user.response.kakao.KakaoTokenResponse;
+import com.moim.backend.domain.user.response.kakao.KakaoUserResponse;
 import com.moim.backend.global.common.exception.CustomException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
