+++ src/main/java/com/moim/backend/domain/user/service/NaverLoginService.java
@@ -3,8 +3,8 @@ package com.moim.backend.domain.user.service;
 import com.moim.backend.domain.user.config.NaverProperties;
 import com.moim.backend.domain.user.config.Platform;
 import com.moim.backend.domain.user.entity.Users;
-import com.moim.backend.domain.user.response.NaverTokenResponse;
-import com.moim.backend.domain.user.response.NaverUserResponse;
+import com.moim.backend.domain.user.response.naver.NaverTokenResponse;
+import com.moim.backend.domain.user.response.naver.NaverUserResponse;
 import com.moim.backend.global.common.exception.CustomException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
